Create README.md
I need all the sponsors I can get. I am working on a dropshipping app that basically makes it as easy as copy and pasting for business owners to make their own websites. Please feel free to reach out if you want to donate or help with the build in any way at all. Thanks alot